+++ streams/src/test/java/org/apache/kafka/streams/state/KeyValueStoreTestDriver.java
@@ -215,14 +215,14 @@ public class KeyValueStoreTestDriver<K, V> {
             new MockStreamsMetrics(new Metrics())
         ) {
             @Override
-            public <K1, V1> int send(final String topic,
-                                     final K1 key,
-                                     final V1 value,
-                                     final Headers headers,
-                                     final Integer partition,
-                                     final Long timestamp,
-                                     final Serializer<K1> keySerializer,
-                                     final Serializer<V1> valueSerializer) {
+            public <K1, V1> long send(final String topic,
+                                      final K1 key,
+                                      final V1 value,
+                                      final Headers headers,
+                                      final Integer partition,
+                                      final Long timestamp,
+                                      final Serializer<K1> keySerializer,
+                                      final Serializer<V1> valueSerializer) {
                 // for byte arrays we need to wrap it for comparison
 
                 final byte[] keyBytes = keySerializer.serialize(topic, headers, key);
@@ -236,14 +236,14 @@ public class KeyValueStoreTestDriver<K, V> {
             }
 
             @Override
-            public <K1, V1> int send(final String topic,
-                                     final K1 key,
-                                     final V1 value,
-                                     final Headers headers,
-                                     final Long timestamp,
-                                     final Serializer<K1> keySerializer,
-                                     final Serializer<V1> valueSerializer,
-                                     final StreamPartitioner<? super K1, ? super V1> partitioner) {
+            public <K1, V1> long send(final String topic,
+                                      final K1 key,
+                                      final V1 value,
+                                      final Headers headers,
+                                      final Long timestamp,
+                                      final Serializer<K1> keySerializer,
+                                      final Serializer<V1> valueSerializer,
+                                      final StreamPartitioner<? super K1, ? super V1> partitioner) {
                 throw new UnsupportedOperationException();
             }
         };
